+++ Heap/heap.c
@@ -22,7 +22,7 @@ struct heap{
 // return 1 when the two floating point numbers are equal, 0 otherwise
 static int equal(float a, float b)
 {
-	if (fabs(a-b) < 1.0e-8){
+	if (fabsf(a-b) < 1.0e-8){
 		return 1 ;
 	}
 	return 0 ;
